cleanup TaikoL1 tests with refactored code to utils folders before attempting tokenomics tests, as code will largely need to be re-used.
