+++ packages/protocol/test/tokenomics/Tokenomics.test.ts
@@ -1,0 +1,3 @@
+describe("tokenomics", function () {
+    it("runs", async function () {});
+});
