+++ packages/protocol/test/utils/block_metadata.ts
@@ -1,0 +1,17 @@
+import { BigNumberish } from "ethers";
+
+type BlockMetadata = {
+    id: number;
+    l1Height: number;
+    l1Hash: string;
+    beneficiary: string;
+    txListHash: string;
+    mixHash: string;
+    extraData: string;
+    gasLimit: BigNumberish;
+    timestamp: number;
+    commitSlot: number;
+    commitHeight: number;
+};
+
+export { BlockMetadata };
